requirements.txt
streamlit
langchain-core
langchain-community
langchain-groq
chromadb
pypdf
python-dotenv